requirements.txt
streamlit>=1.33
pandas>=2.0
pyyaml>=6.0
openpyxl>=3.1
plotly>=5.0
transformers>=4.40
torch
accelerate
